+++ demos/employees/src/main/java/is/codion/framework/demos/employees/model/EmployeeEditModel.java
@@ -69,7 +69,7 @@ public final class EmployeeEditModel extends SwingEntityEditModel {
 			//only show managers from the same department as the selected employee and hide the currently
 			//selected employee to prevent an employee from being made her own manager
 			foreignKeyComboBoxModel(Employee.MANAGER_FK).includeCondition().set(manager ->
-							Objects.equals(manager.referencedEntity(Employee.DEPARTMENT_FK), department)
+							Objects.equals(manager.entity(Employee.DEPARTMENT_FK), department)
 											&& !Objects.equals(manager, entity()));
 		});
 	}