+++ framework/domain/src/test/java/is/codion/framework/domain/entity/EntitySerializerTest.java
@@ -101,12 +101,12 @@ public final class EntitySerializerTest {
 		serializeDeserialize(serializer, serializer, entity, deserializedEntity);
 		assertTrue(deserializedEntity.modified(Employee.NAME));
 		assertTrue(Entity.valuesEqual(entity, deserializedEntity));
-		assertTrue(Entity.valuesEqual(entity.referencedEntity(Employee.DEPARTMENT_FK), deserializedEntity.referencedEntity(Employee.DEPARTMENT_FK)));
+		assertTrue(Entity.valuesEqual(entity.entity(Employee.DEPARTMENT_FK), deserializedEntity.entity(Employee.DEPARTMENT_FK)));
 		assertFalse(deserializedEntity.get(Employee.DEPARTMENT_FK).mutable());
-		Entity manager = entity.referencedEntity(Employee.MANAGER_FK);
-		Entity deserializedManager = deserializedEntity.referencedEntity(Employee.MANAGER_FK);
+		Entity manager = entity.entity(Employee.MANAGER_FK);
+		Entity deserializedManager = deserializedEntity.entity(Employee.MANAGER_FK);
 		assertTrue(Entity.valuesEqual(manager, deserializedManager));
-		assertTrue(Entity.valuesEqual(manager.referencedEntity(Employee.DEPARTMENT_FK), deserializedManager.referencedEntity(Employee.DEPARTMENT_FK)));
+		assertTrue(Entity.valuesEqual(manager.entity(Employee.DEPARTMENT_FK), deserializedManager.entity(Employee.DEPARTMENT_FK)));
 
 		DefaultKey key = createTestKey();
 		DefaultKey deserializedKey = (DefaultKey) ENTITIES.keyBuilder(CompositeMaster.TYPE).build();