+++ framework/json-domain/src/test/java/is/codion/framework/json/domain/EntityObjectMapperTest.java
@@ -117,7 +117,7 @@ public final class EntityObjectMapperTest {
 		Entity readEmp = mapper.readValue(jsonString, Entity.class);
 		assertTrue(emp.valuesEqual(readEmp));
 
-		Entity readDept = readEmp.referencedEntity(Employee.DEPARTMENT_FK);
+		Entity readDept = readEmp.entity(Employee.DEPARTMENT_FK);
 		assertTrue(dept.valuesEqual(readDept));
 	}
 
@@ -222,8 +222,8 @@ public final class EntityObjectMapperTest {
 		jsonString = mapper.writeValueAsString(singletonList(emp1));
 		Entity emp1Deserialized = mapper.deserializeEntities(jsonString).get(0);
 		assertTrue(emp1.valuesEqual(emp1Deserialized));
-		assertTrue(emp1.referencedEntity(Employee.DEPARTMENT_FK).valuesEqual(emp1Deserialized.referencedEntity(Employee.DEPARTMENT_FK)));
-		assertTrue(emp1.referencedEntity(Employee.MGR_FK).valuesEqual(emp1Deserialized.referencedEntity(Employee.MGR_FK)));
+		assertTrue(emp1.entity(Employee.DEPARTMENT_FK).valuesEqual(emp1Deserialized.entity(Employee.DEPARTMENT_FK)));
+		assertTrue(emp1.entity(Employee.MGR_FK).valuesEqual(emp1Deserialized.entity(Employee.MGR_FK)));
 
 		LocalDate newHiredate = LocalDate.parse("2002-11-21", format);
 		emp1.put(Employee.COMMISSION, 550.55);